# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/rain": {
            "get": {
                "description": "get rain status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rain"
                ],
                "summary": "Get rain status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
        "/weather/{city}": {
            "get": {
                "description": "get weather by city",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "weather"
                ],
                "summary": "Get weather by city",
                "parameters": [
                    {
                        "type": "string",
                        "description": "City",
                        "name": "city",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Weather"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "httputil.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "model.Weather": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "object",
                    "properties": {
                        "cloud": {
                            "type": "integer"
                        },
                        "condition": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "icon": {
                                    "type": "string"
                                },
                                "text": {
                                    "type": "string"
                                }
                            }
                        },
                        "feelslike_c": {
                            "type": "number"
                        },
                        "feelslike_f": {
                            "type": "number"
                        },
                        "gust_kph": {
                            "type": "number"
                        },
                        "gust_mph": {
                            "type": "number"
                        },
                        "humidity": {
                            "type": "integer"
                        },
                        "is_day": {
                            "type": "integer"
                        },
                        "last_updated": {
                            "type": "string"
                        },
                        "last_updated_epoch": {
                            "type": "integer"
                        },
                        "precip_in": {
                            "type": "number"
                        },
                        "precip_mm": {
                            "type": "number"
                        },
                        "pressure_in": {
                            "type": "number"
                        },
                        "pressure_mb": {
                            "type": "number"
                        },
                        "temp_c": {
                            "type": "number"
                        },
                        "temp_f": {
                            "type": "number"
                        },
                        "uv": {
                            "type": "number"
                        },
                        "vis_km": {
                            "type": "number"
                        },
                        "vis_miles": {
                            "type": "number"
                        },
                        "wind_degree": {
                            "type": "integer"
                        },
                        "wind_dir": {
                            "type": "string"
                        },
                        "wind_kph": {
                            "type": "number"
                        },
                        "wind_mph": {
                            "type": "number"
                        }
                    }
                },
                "location": {
                    "type": "object",
                    "properties": {
                        "country": {
                            "type": "string"
                        },
                        "lat": {
                            "type": "number"
                        },
                        "localtime": {
                            "type": "string"
                        },
                        "localtime_epoch": {
                            "type": "integer"
                        },
                        "lon": {
                            "type": "number"
                        },
                        "name": {
                            "type": "string"
                        },
                        "region": {
                            "type": "string"
                        },
                        "tz_id": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}